src/passwordmanager/requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
